Add files via upload
Assignment 1, Innopolis University
Intro to ML, Fall 2022 - Bachelors

In this assignment, I was required to implement and compare various regression and classification techniques. The task is divided into two parts: practical
task and theoretical part. In the practical task, I solved a classification task where I implemented regression techniques in the initial phase, the details will be explained the Section 1; the Section 2 is about the theoretical part, where I answered the questions related to the work that I did.